MissionPage (WIP)
Missions = YES
Badges = (WIP)

+++ app/src/main/java/com/example/alpvisualprogramming/model/Badge.kt
@@ -1,9 +1,9 @@
 package com.example.alpvisualprogramming.model
 
 data class Badge(
-    var id: Int = 0,
+    var id: Double = 0.0,
     var image: String = "",
     var name: String = "",
-    var price: Int = 0,
+    var price: Double = 0.0,
 //    var user_id: String = "",
 )
